Update sheets conn and move out of scripts folder
