+++ .env-sample
@@ -16,8 +16,3 @@ ACTIVE_ROLE =
 
 Paste your MongoDB URI below.
 DATABASE_URI =
-
-Paste your Google Sheets Project Key below.
-GOOGLE_EMAIL = 
-GOOGLE_PRIVATE_KEY = 
-GOOGLE_SPREADSHEET_ID =
