+++ src/connections/google-sheets-conn.js
@@ -1,8 +1,9 @@
 const { google } = require("googleapis");
+const keys = require("../../google-sheets.json");
 // Declare scopes that determine the authorization level
 const googleScopes = ["https://www.googleapis.com/auth/spreadsheets"];
 // Create client
-const gClient = new google.auth.JWT(process.env.GOOGLE_EMAIL, null, process.env.GOOGLE_PRIVATE_KEY, googleScopes);
+const gClient = new google.auth.JWT(keys.client_email, null, keys.private_key, googleScopes);
 
 // Connect to Google Sheets
 gClient.authorize((err, tokens) => {
@@ -13,10 +14,10 @@ gClient.authorize((err, tokens) => {
 	}
 });
 
-async function getVocab(range) {
+async function fetchVocab(range) {
 	const gSheetsAPI = google.sheets({ version: "v4", auth: gClient });
 	const request = {
-		spreadsheetId: process.env.GOOGLE_SPREADSHEET_ID,
+		spreadsheetId: keys.spreadsheet_id,
 		range: range,
 	};
 
@@ -33,4 +34,4 @@ function parseVocab(result) {
 	}));
 }
 
-module.exports = { getVocab };
+module.exports = { fetchVocab };
